feat: 딜러 수익 계산 및 출력 로직 구현

+++ src/main/java/blackjack/model/results/Result.java
@@ -3,34 +3,34 @@ package blackjack.model.results;
 public enum Result {
     WIN_BY_BLACKJACK {
         @Override
-        public double calculate(int money) {
-            return (money * 1.5);
+        public int calculate(int money) {
+            return (int) (money * 1.5);
         }
     },
     WIN {
         @Override
-        public double calculate(int money) {
+        public int calculate(int money) {
             return money;
         }
     },
     PUSH {
         @Override
-        public double calculate(int money) {
+        public int calculate(int money) {
             return 0;
         }
     },
     LOSE {
         @Override
-        public double calculate(int money) {
+        public int calculate(int money) {
             return -money;
         }
     },
     LOSE_BY_BLACKJACK {
         @Override
-        public double calculate(int money) {
-            return -(money * 1.5);
+        public int calculate(int money) {
+            return (int) -(money * 1.5);
         }
     };
 
-    public abstract double calculate(int money);
+    public abstract int calculate(int money);
 }
